+++ home.php
@@ -28,7 +28,6 @@
       }
 
       if (is_front_page() && wrdsb_i_am_a_school()) { ?>
-        <p><a href="http://bit.ly/2gy8HyY" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'wearewrdsb');"><img src="https://www.wrdsb.ca/wp-content/uploads/WRDSB-CTA.png" alt="Vote for the new purpose statement before December 11, 2016 #WeAreWRDSB"/></a></p>
        <p><a href="https://myway.wrdsb.ca/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'myway'"><img src="https://www.wrdsb.ca/wp-content/uploads/myway_banner_344x100.jpg" alt="MyWay Logo"/></a></p>
         <p><a href="<?php echo site_url();?>/about/school-year-information" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'syi', '/about/school-year-information',{'nonInteraction':1});"><img src="https://www.wrdsb.ca/wp-content/uploads/schoolyearinformation_344x100.jpg" alt="School Year Information Logo"/></a></p>
         <p><a href="https://www.canadahelps.org/dn/15506" target="_blank" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'wefi', 'https://www.canadahelps.org/dn/15506',{'nonInteraction':1});"><img src="https://www.wrdsb.ca/wefi/wp-content/uploads/sites/19/wefi_donate.jpg" alt="Donate Online! Waterloo Education Foundation Inc. (WEFI)"></a></p>
@@ -66,8 +65,7 @@
       }
 
       if (is_front_page() && wrdsb_i_am_a_school()) { ?>
-        <p><a href="http://bit.ly/2gy8HyY" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'wearewrdsb');"><img src="https://www.wrdsb.ca/wp-content/uploads/WRDSB-CTA.png" alt="Vote for the new purpose statement before December 11, 2016 #WeAreWRDSB"/></a></p>
-       <p><a href="https://myway.wrdsb.ca/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'myway'"><img src="https://www.wrdsb.ca/wp-content/uploads/myway_banner_344x100.jpg" alt="MyWay Logo"/></a></p>
+        <p><a href="https://myway.wrdsb.ca/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'myway'"><img src="https://www.wrdsb.ca/wp-content/uploads/myway_banner_344x100.jpg" alt="MyWay Logo"/></a></p>
         <p><a href="/about/school-year-information" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'syi');"><img src="https://www.wrdsb.ca/wp-content/uploads/schoolyearinformation_344x100.jpg" alt="School Year Information Logo"/></a></p>
         <p><a href="https://www.canadahelps.org/dn/15506" target="_blank" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'wefi', 'https://www.canadahelps.org/dn/15506',{'nonInteraction':1});"><img src="https://www.wrdsb.ca/wefi/wp-content/uploads/sites/19/wefi_donate.jpg" alt="Donate Online! Waterloo Education Foundation Inc. (WEFI)"></a></p>
         <iframe id="school-day" src="https://www.school-day.com/pg/school/sso/index.php"></iframe>